services/ml_inference/requirements.txt
kafka-python
psycopg2-binary
python-dotenv
pandas
scikit-learn
joblib

